refactor: divide setting column part and storage part
